Moved Loot Disk Cycling handler into separate file.
