+++ src/main/scala/li/cil/oc/integration/jei/ModPluginOpenComputers.scala
@@ -1,12 +1,6 @@
 package li.cil.oc.integration.jei
 
-import java.util
-
-import li.cil.oc.Constants
 import li.cil.oc.Settings
-import li.cil.oc.api
-import li.cil.oc.common.Loot
-import li.cil.oc.common.recipe.LootDiskCyclingRecipe
 import li.cil.oc.integration.util.ItemBlacklist
 import mezz.jei.api.IItemRegistry
 import mezz.jei.api.IJeiHelpers
@@ -15,14 +9,6 @@ import mezz.jei.api.IModPlugin
 import mezz.jei.api.IModRegistry
 import mezz.jei.api.IRecipeRegistry
 import mezz.jei.api.JEIPlugin
-import mezz.jei.api.recipe.BlankRecipeWrapper
-import mezz.jei.api.recipe.IRecipeHandler
-import mezz.jei.api.recipe.IRecipeWrapper
-import mezz.jei.api.recipe.VanillaRecipeCategoryUid
-import mezz.jei.api.recipe.wrapper.ICraftingRecipeWrapper
-import net.minecraft.item.ItemStack
-
-import scala.collection.convert.WrapAsJava._
 
 @JEIPlugin
 class ModPluginOpenComputers extends IModPlugin {
@@ -46,21 +32,4 @@ class ModPluginOpenComputers extends IModPlugin {
 
   override def onRuntimeAvailable(jeiRuntime: IJeiRuntime): Unit = {
   }
-
-  object LootDiskCyclingRecipeHandler extends IRecipeHandler[LootDiskCyclingRecipe] {
-    override def getRecipeClass: Class[LootDiskCyclingRecipe] = classOf[LootDiskCyclingRecipe]
-
-    override def getRecipeCategoryUid: String = VanillaRecipeCategoryUid.CRAFTING
-
-    override def getRecipeWrapper(recipe: LootDiskCyclingRecipe): IRecipeWrapper = new LootDiskCyclingRecipeWrapper(recipe)
-
-    override def isRecipeValid(recipe: LootDiskCyclingRecipe): Boolean = true
-  }
-
-  class LootDiskCyclingRecipeWrapper(val recipe: LootDiskCyclingRecipe) extends BlankRecipeWrapper with ICraftingRecipeWrapper {
-    override def getInputs: util.List[_] = List(seqAsJavaList(Loot.worldDisks.map(_._1)), api.Items.get(Constants.ItemName.Wrench).createItemStack(1))
-
-    override def getOutputs: util.List[ItemStack] = Loot.worldDisks.map(_._1).toList
-  }
-
 }
